+++ tests/unittests/test_config.py
@@ -1,7 +1,7 @@
-from pyqgiswps.config import get_size_bytes
 
 
 def test_size_bytes():
+    from pyqgiswps.config import get_size_bytes
 
     sz = get_size_bytes('1m')
     assert sz == 1048576
@@ -11,3 +11,13 @@ def test_size_bytes():
 
     sz = get_size_bytes('2g')
     assert sz == 2147483648
+
+
+def test_manifest():
+    from pyqgiswps import version
+
+    m = version.read_manifest()
+    print("\n::test_manifest::", m)
+    assert 'commitid' in m
+    assert 'buildid' in m
+    assert 'version' in m